+++ src/App.tsx
@@ -10,15 +10,39 @@ import HomePage from "./pages/HomePage";
 import Header from "./components/Header";
 import { Routes, Route } from "react-router-dom";
 import AddOnPage from "./pages/AddOnPage";
-import { EventBus } from "./components/EventBus";
+import { useAppSelector } from "./app/hooks";
+import { RootState } from "./app/store";
+import { useAuth } from "./features/authentication/useAuth";
+import { useEffect } from "react";
+import { Broker } from "./broker/broker";
 
 /**
  * The central application component, responsible for high-level page layout and routing.
  */
 const App = () => {
+  const auth = useAppSelector((state: RootState) => state.auth);
+  const { login } = useAuth();
+
+  useEffect(() => {
+    login();
+  }, []);
+
+  // Connects the WebSocket and sets authorisation header for the broker
+  useEffect(() => {
+    if (auth.authorized && auth.jwt) {
+      console.log("Connecting broker");
+      Broker.instance()
+        .useAuth(auth)
+        .connect(() => {
+          console.log("WS connected", window.location.search);
+        });
+    } else {
+      // dispatch(logout());
+    }
+  }, [auth]);
+
   return (
     <>
-      <EventBus />
       <Header />
       <Routes>
         <Route path="/" element={<HomePage />} />
